Add product field to Plan object.
According to https://stripe.com/docs/api/plans, a plan object always
contains a field `product` containing the id of the product the plan is
a part of.

This commit adds the `product` field to the Plan struct, with the goal
of enabling simpler handling of subscription flows when using webhooks.